fixed compilation warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/fraqtive/code/trunk@65 21e5ddf2-8f81-477f-a659-84387f25dd0c

+++ src/bookmarkmodel.cpp
@@ -262,7 +262,7 @@ void BookmarkModel::calculate( const Bookmark& bookmark, double* buffer, const Q
 
     GeneratorSettings settings = DataFunctions::defaultGeneratorSettings();
 
-    int maxIterations = pow( 10.0, settings.calculationDepth() ) * qMax( 1.0, 1.45 + position.zoomFactor() );
+    int maxIterations = (int)( pow( 10.0, settings.calculationDepth() ) * qMax( 1.0, 1.45 + position.zoomFactor() ) );
     double threshold = settings.detailThreshold();
 
 #if defined( HAVE_SSE2 )
